moved db related files together
