+++ server/server.js
@@ -1,5 +1,5 @@
 const app = require("./app");
-const db = require("./db");
+const db = require("./src/db");
 const port = 5001;
 
 db.connect().then(() => {
